Fix sleep timer and stop using System.exit
- Fix sleep timer
The wait method would always sleep for 0 ms instead of the time of the parameter. This was originally for debugging purposes, but was not changed back.

- Change all methods and Game class to non-static
Since the Main class called the method to start the game directly on the Game class, the methods could not be static. Once the Main class created an object of class Game, the methods could be written as static.

- Add comments to the methods missing them

- Stop using System.exit to end the game
System.exit(0) is an easy way to quit the game, but is not recommended. By using a boolean value gameActive, to tell the program at certain points to do nothing as opposed to running as normal, the program can finish execution.